creating logical itens and cadaster itens and correction all sbs

+++ src/main/java/HexTech/Backend_lV_Fatec_Embraer/item/services/ListItemsByChassi/ListItemsByChassiController.java
@@ -22,7 +22,7 @@ public class ListItemsByChassiController {
 	
 	@CrossOrigin
 	@GetMapping
-	public ResponseEntity<ListItemsResponseDTO> listItemsByChassy(@PathVariable("chassi") Integer chassi) {
+	public ResponseEntity<ListItemsResponseDTO> listItemsByChassy(@PathVariable("chassi") Long chassi) {
 		
 		ListItemsResponseDTO listItems = itemsByChassiService.execute(chassi);
